ツリー外からのドロップを無視
git-svn-id: https://www.cozmixng.org/repos/piro/bookmarks2pane/trunk@2134 599a83e7-65a4-db11-8015-0010dcdd6dc2

+++ content/bookmarks2pane/bookmarksPanelOverlay.js
@@ -530,6 +530,9 @@ var Bookmarks2PaneService = {
 	dustboxDNDObserver : {
 		onDrop : function(aEvent, aTransferData, aSession)
 		{
+			if (!aSession.sourceNode ||
+				aSession.sourceNode.parentNode != Bookmarks2PaneService.currentTree)
+				return;
 			Bookmarks2PaneService.deleteCurrentSelection();
 		},
 		onDragOver : function() {},
